refactor: heading and paragraph
