+++ apps/frontend/components/Layout/index.ts
@@ -1,1 +1,3 @@
+export { default as Header } from './Header';
 export { default as PageContent } from './PageContent';
+export { default as Paragraph } from './Paragraph';
